% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{build_keyboard_df}
\alias{build_keyboard_df}
\title{Build keyboard data frame}
\usage{
build_keyboard_df(candidates, answer)
}
\arguments{
\item{candidates}{A vector of candidate words}

\item{answer}{The answer word}
}
\value{

}
\description{
Build keyboard data frame
}
