% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{build_puzzle_df}
\alias{build_puzzle_df}
\title{Build a data frame for plotting the puzzle grid}
\usage{
build_puzzle_df(candidates, answer, total_try = 6)
}
\arguments{
\item{candidates}{A vector of candidate words}

\item{answer}{The answer word}

\item{total_try}{Total number of tries allowed.}
}
\value{
A data frame containing `word`, `letter`, `letter_idx` and `clue` columns
}
\description{
Build a data frame for plotting the puzzle grid
}
