% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordle_check.R
\name{check_full_game}
\alias{check_full_game}
\title{Identify remaining words after multiple guesses for a single target word}
\usage{
check_full_game(dictionary, target, guesses, hide = TRUE)
}
\arguments{
\item{dictionary}{Character vector of words to filter based on result of comparing \code{guess} to \code{target}}

\item{target}{Target of the match}

\item{guesses}{Character vector of guesses (in order) to compare to target word}

\item{hide}{If \code{TRUE}, return list of integers indicating number of words
remaining in \code{dictionary} after each guess is compared to the target word. Otherwise,
return list of Character vectors of the words that remain.}
}
\value{
Either list of integers (if \code{hide = TRUE}) or character vectors (if \code{hide = FALSE})
}
\description{
Given a provided dictionary, compare multiple guesses to a single
target word, returning the remaining available words in the dictionary at each
step based on the correct, incorrect, and misplaced letters identified after
comparing each guess to the target word.
}
\examples{
all_words <- words::words$word
dictionary <- all_words[stringr::str_count(all_words) == 5]

target <- "tangy"
guesses <- c("plaid", "cream", "axons", "tangy")

check_full_game(dictionary, target, guesses, hide = FALSE)
}
