% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_wordle.R
\name{check_guess}
\alias{check_guess}
\title{Compare user input to target word}
\usage{
check_guess(target, n_letters, init = TRUE)
}
\arguments{
\item{target}{Target word of the game}

\item{n_letters}{Number of letters of the words being used in this game}

\item{init}{Whether this is the initial guess of the game}
}
\description{
Prompt user for input and provide feedback. Continue to do so until
game is won or exited.
}
