% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordle_check.R
\name{check_remaining}
\alias{check_remaining}
\title{Determine possible remaining words based on results of a Wordle guess}
\usage{
check_remaining(dictionary, target, guess, hide = TRUE)
}
\arguments{
\item{dictionary}{Character vector of words to filter based on result of comparing \code{guess} to \code{target}}

\item{target}{target of the match}

\item{guess}{word to be compared to the target word}

\item{hide}{If \code{TRUE}, return number of words remaining in dictionary after comparing
\code{target} to \code{guess}. Otherwise, return character vector of remaining words themselves}
}
\value{
If \code{hide = TRUE}, single integer. Otherwise, character vector
}
\description{
Filter provided dictionary based on the logic resulting from
feedback after a single wordle guess.
}
\examples{
all_words <- words::words$word
dictionary <- all_words[stringr::str_count(all_words) == 5]

check_remaining(dictionary, "torch", "trick", hide = TRUE)
}
