% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordle.R
\name{collect_colors}
\alias{collect_colors}
\title{Extract structured vectors from colors string}
\usage{
collect_colors(colors)
}
\arguments{
\item{colors}{String of length 5, G = green, Y = yellow, '-' = gray}
}
\value{
Named list of 3 logical vectors indicating color
}
\description{
Extract structured vectors from colors string
}
\examples{
collect_colors('-YY-G')
}
