% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compare_word}
\alias{compare_word}
\title{Compare candidate letter with the answer}
\usage{
compare_word(candidate, answer)
}
\arguments{
\item{candidate}{A candidate word}

\item{answer}{The answer word}
}
\value{
A numeric vector.
}
\description{
Compare candidate letter with the answer
}
\details{
Returns a vector of values
 0 = "not in the answer"
 1 = "in the answer but incorrect position"
 2 = "correct position in the answer"
}
