% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary}
\alias{dictionary}
\title{Wordle dictionary}
\format{
Character vector
}
\source{
\url{https://www.powerlanguage.co.uk/}
}
\usage{
dictionary
}
\description{
12,972 5-letter words
}
\keyword{datasets}
