% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dictionary.R
\name{filter_correct}
\alias{filter_correct}
\title{Filter dictionary of target words based on letters that are correct}
\usage{
filter_correct(dictionary, letters, positions)
}
\arguments{
\item{dictionary}{Character vector of possible words to match}

\item{letters}{List of single-length character strings, indicating the letters
that are present in the word.}

\item{positions}{List of integers between 1 and 5 that indicates the index within
the word where each associated (at same list index) correct letter was guessed.}
}
\value{
Character vector of possible target words remaining after applying correct letter logic
}
\description{
Filter dictionary of target words based on letters that are correct
}
\examples{
dictionary <- c("cumin", "salad", "touch", "urges")
filter_correct(dictionary, letters = list("o", "h"), positions = list(2, 5))
}
