% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dictionary.R
\name{filter_game}
\alias{filter_game}
\title{Filter dictionary of target words using Wordle logic}
\usage{
filter_game(guesses, target, dict = NULL, hide = TRUE)
}
\arguments{
\item{guesses}{vector of character indicating sequential guesses}

\item{target}{Word to use as comparison to \code{guess} to determine
which letters are correct, incorrect, or misplaced.}

\item{dict}{Vector of strings of possible words to match. If NULL, uses \code{dictionary} as default.}

\item{hide}{If \code{TRUE}, return number of words remaining after filtering. If
\code{FALSE}, return vector of remaining words themselves.}
}
\value{
List of either a vector of words remaining after filtering or integer
indicating the number of words remaining.
}
\description{
Restrict original dictionary of possible words to those that meet criteria
based on information about letters that are correct, incorrect, or misplaced
in the unknown target word.
}
\examples{
target <- "panic"
guess <- c("burst", "canoe")

filter_game(dict = dictionary, guess, target)
}
