% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordle.R
\name{filter_grays}
\alias{filter_grays}
\title{Filter word list to exclude those with gray letters}
\usage{
filter_grays(words, word, colors)
}
\arguments{
\item{words}{Input word vector}

\item{word}{Word to process}

\item{colors}{Output from \code{\link[=collect_colors]{collect_colors()}} for this word}
}
\value{
Output word vector
}
\description{
Filter word list to exclude those with gray letters
}
