% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dictionary.R
\name{filter_guess}
\alias{filter_guess}
\title{Filter dictionary of target words using Wordle logic}
\usage{
filter_guess(
  dictionary,
  guess,
  target = NULL,
  correct = NULL,
  incorrect = NULL,
  misplaced = NULL,
  hide = TRUE
)
}
\arguments{
\item{dictionary}{Vector of strings of possible words to match}

\item{guess}{String of guessed word}

\item{target}{If provided, word to use as comparison to \code{guess} to determine
which letters are correct, incorrect, misplaced. If not specified, at least
one of correct, incorrect, or misplaced must be specified.}

\item{correct}{Vector of integers indicating the index position of all correct
letters in \code{guess}}

\item{incorrect}{Vector of integers indicating the index position of all incorrect
letters in \code{guess}}

\item{misplaced}{Vector of integers indicating the index position of all misplaced
letters in \code{guess}}

\item{hide}{If \code{TRUE}, return number of words remaining after filtering. If
\code{FALSE}, return vector of remaining words themselves.}
}
\value{
Either a vector of words remaining after filtering or integer
indicating the number of words remaining.
}
\description{
Restrict original dictionary of possible words to those that meet criteria
based on information about letters that are correct, incorrect, or misplaced
in the unknown target word.
}
\examples{
all_words <- words::words$word
dictionary <- all_words[stringr::str_count(all_words) == 5]

target <- "tangy"
guess <- "tonal"

filter_guess(dictionary, guess, target)

filter_guess(
  dictionary,
  guess,
  correct = c(1, 3),
  incorrect = c(2, 5),
  misplaced = 4
)
}
