% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dictionary.R
\name{filter_incorrect}
\alias{filter_incorrect}
\title{Filter dictionary of target words based on letters that are incorrect}
\usage{
filter_incorrect(dictionary, letters)
}
\arguments{
\item{dictionary}{Character vector of possible words to match}

\item{letters}{vector of single-length strings indicating letters not present in target word.}
}
\value{
Character vector of possible target words remaining after applying incorrect letter logic
}
\description{
Filter dictionary of target words based on letters that are incorrect
}
\examples{
dictionary <- c("cumin", "salad", "touch", "urges")
filter_incorrect(dictionary, letters = list("t", "e"))
}
