% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dictionary.R
\name{filter_misplaced}
\alias{filter_misplaced}
\title{Filter dictionary of target words based on letters that are misplaced.}
\usage{
filter_misplaced(dictionary, letters, positions)
}
\arguments{
\item{dictionary}{Character vector of possible words to match}

\item{letters}{List of single-length character strings, indicating the letters
that are misplaced.}

\item{positions}{List of integers between 1 and 5 that indicates the index within
the word where each associated (at same list index) misplaced letter was guessed.}
}
\value{
Character vector of possible target words remaining after applying misplaced letter logic
}
\description{
Filter dictionary of target words based on letters that are misplaced.
}
\examples{
dictionary <- c("cumin", "salad", "touch", "urges")
filter_misplaced(dictionary, letters = list("a", "l"), positions = list(1, 2))
}
