% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dictionary.R
\name{filter_wordle}
\alias{filter_wordle}
\alias{filter_correct}
\alias{filter_incorrect}
\alias{filter_misplaced}
\title{Filter dictionary of target words using Wordle logic}
\usage{
filter_correct(letters, positions, dict = NULL)

filter_incorrect(letters, dict = NULL)

filter_misplaced(letters, positions, dict = NULL)
}
\arguments{
\item{letters}{Vector of single-length character strings, indicating the letters
that are correct, incorrect, or misplaced.}

\item{positions}{For \code{filter_correct} and \code{filter_misplaced},
vector of integers between 1 and 5 that indicates the index within
the word where each associated misplaced letter was guessed.}

\item{dict}{Vector of strings of possible words to match}
}
\value{
Character vector of words in \code{dictionary} that meet specified criteria
}
\description{
Restrict original dictionary of possible words to those that meet criteria
based on information about letters that are correct, incorrect, or misplaced
in the unknown target word.
}
