% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordle.R
\name{filter_words}
\alias{filter_words}
\title{Filter words based upon "Wordle" game state}
\usage{
filter_words(
  words,
  exact = ".....",
  excluded_letters = "",
  wrong_spot = c("", "", "", "", ""),
  known_count = c()
)
}
\arguments{
\item{words}{character vector of candidate words}

\item{exact}{single string representing known characters in the word, with
'.' used to indicate the letter at this position is unknown.
E.g. For a 5 letter word if the 3rd and 4th letters are known to be 'a' and 'c', but
all other letters are unknown, then \code{exact = "..ac."}}

\item{excluded_letters}{string containing letters known to not be in the word}

\item{wrong_spot}{a character vector the same length as the number of letters
in the target word.  Each string in this vector represents all letters
which are known to be part of the word, but in the wrong spot.
E.g. if 'a' has been attempted as the first character, and it exists
in the word, but worlde claims it is not yet in the correct position,
then \code{wrong_spot = c('a', '', '', '', '')}}

\item{known_count}{name character vector giving letters and their known counts.
E.g. This can be used if it is known definitively that there is only
one letter 'e' in the target word, in which case
\code{known_count = c(e = 1)}}
}
\description{
Filter words based upon "Wordle" game state
}
