% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_letters.R
\name{find_letters}
\alias{find_letters}
\alias{find_correct}
\alias{find_incorrect}
\alias{find_misplaced}
\title{Determine letter categories for a Wordle guess}
\usage{
find_correct(target, guess)

find_incorrect(target, guess)

find_misplaced(target, guess)
}
\arguments{
\item{target}{target of the match}

\item{guess}{word to be compared to the target word}
}
\value{
integer vector
}
\description{
Compare a guess word to a known target word and determine the index locations
for all correct, incorrect, and misplaced letters in the guess. Guess word
and target word must have the same number of letters.
}
