% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_letters.R
\name{find_misplaced}
\alias{find_misplaced}
\alias{find_incorrect}
\alias{find_correct}
\title{Identify correct, incorrect, and misplaced letters for a given target and guess word}
\usage{
find_misplaced(target, guess, idx = FALSE)

find_incorrect(target, guess, idx = FALSE)

find_correct(target, guess, idx = FALSE)
}
\arguments{
\item{target}{target of the match}

\item{guess}{word to be compared to the target word}

\item{idx}{If \code{TRUE} return integer index of position of correct, incorrect, or misplaced letters.
If \code{FALSE} return letter characters themselves.}
}
\value{
integer or character vector
}
\description{
Identify correct, incorrect, and misplaced letters for a given target and guess word
}
\examples{
find_misplaced("torch", "trick")
find_correct("torch", "trick")
}
