% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dictionary.R
\name{find_words}
\alias{find_words}
\title{Filter dictionary using regular expression}
\usage{
find_words(pattern, dict = NULL, neg = FALSE)
}
\arguments{
\item{pattern}{Regular expression to match with entries in \code{dict}}

\item{dict}{Vector of words through which to search for pattern}

\item{neg}{If \code{TRUE}, include strings in \code{dict} that match with \code{pattern}. Otherwise, exclude matches.}
}
\value{
Character vector
}
\description{
Filter dictionary using regular expression
}
