% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_state.R
\name{generate_available_cols}
\alias{generate_available_cols}
\title{Determine which columns make sense to guess}
\usage{
generate_available_cols(game_state, verbose = TRUE)
}
\arguments{
\item{game_state}{The current game_state}
}
\value{
A logical vector of logical columns to guess
}
\description{
Determine which columns make sense to guess
}
