% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{guess_dictionary}
\alias{guess_dictionary}
\title{Wordle dictionary}
\format{
Character vector
}
\source{
\url{https://www.powerlanguage.co.uk/}
}
\usage{
guess_dictionary
}
\description{
10,657 5-letter words used to validate user guesses when playing Wordle
}
\keyword{datasets}
