% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/words.R
\docType{data}
\name{letter_freq}
\alias{letter_freq}
\title{Letter frequency ordering in \code{wordle_dict}}
\format{
An object of class \code{character} of length 26.
}
\usage{
letter_freq
}
\description{
Letter frequency ordering in \code{wordle_dict}
}
\keyword{datasets}
