% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_setup.R
\name{load_wordle_word_lists}
\alias{load_wordle_word_lists}
\title{Load the possible word lists from the Wordle source}
\usage{
load_wordle_word_lists(
  url = "https://www.nytimes.com/games/wordle/main.bd4cb59c.js",
  match_possible_words = "balmy",
  match_all_words = "aahed",
  expected_characters = 5
)
}
\arguments{
\item{url}{The javascript source for finding the word lists}

\item{match_possible_words}{A word that is within the word list for the
possible words (mainly, words that people are likely to know, the Wordle
developer was kind)}

\item{match_all_words}{A word that is within the word list for all words
(more than just the possible words)}

\item{expected_characters}{The number of characters expected in every word,
used as a quality check for the outputs}
}
\value{
A named list with names of "possible" and "all" for the possible
and all word lists
}
\description{
Load the possible word lists from the Wordle source
}
