% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_letters.R
\name{make_guess}
\alias{make_guess}
\title{Make a single Wordle guess}
\usage{
make_guess(target, guess)
}
\arguments{
\item{target}{target of the match}

\item{guess}{word to be compared to the target word}
}
\value{
List of length 3. Each list element is a vector of integers indicating
index positions for the letters in the guess word in the associated category.
}
\description{
Compare a guess word to a target word and return index positions
of correct, incorrect, and misplaced letters.
}
\examples{
guess <- make_guess("torch", "trick")
str(guess)
}
