% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_state.R
\name{new_wordle_game}
\alias{new_wordle_game}
\title{Generate a new Wordle game with a correct word}
\usage{
new_wordle_game(correct_word, possible_df = wordle_colors, expected_char = 5)
}
\arguments{
\item{correct_word}{The word that is the correct word for the game}

\item{possible_df}{The data.frame of possible words}

\item{expected_char}{Number of characters expected for the correct_word}
}
\value{
A "wordle_game" object which is a list with values of "correct",
"possible_df", "available_rows" (the rows which have a chance of being
correct as a logical vector), and "available_cols" (the columns which give
non-degenerate information as a logical vector).
}
\description{
Generate a new Wordle game with a correct word
}
