% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthogonal-words.R
\docType{data}
\name{orthogonal_words}
\alias{orthogonal_words}
\title{Orthogonal word sets}
\format{
An object of class \code{list} of length 5.
}
\usage{
orthogonal_words
}
\description{
This is a list of 5 data.frames, with each data.frame representing all
possible word sets of a given length - with the constraint that each set of
words must contain the 'n' most common letters in the wordle dictionary.
}
\details{
The first data.frame contains only words made up of the 5 most
common letters \code{c("s", "e", "a", "o", "r")}

The second data.frame contains two words per row.  Those two words must
contain the first 10 most common letters in the wordle dictionary.
i.e. \code{c("s", "e", "a", "o", "r", "i", "l", "t", "n", "u")}

The third data.frame contains three words per row.  Those three words must
contain the first 15 most common letters in the wordle dictionary.
i.e. \code{c("s", "e", "a", "o", "r", "i", "l", "t", "n", "u", "d", "y",
"c", "p", "m")}

The fourth data.frame contains four words per row.  Those three words must
contain the first 20 most common letters in the wordle dictionary.
i.e. \code{c("s", "e", "a", "o", "r", "i", "l", "t", "n", "u", "d", "y",
"c", "p", "m", "h", "g", "b", "k", "f")}

The fifth data.frame contains five words per row. These five words consist
of 25 distinct letters (no repeats).
}
\keyword{datasets}
