% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_letters.R
\name{parse_word}
\alias{parse_word}
\title{Split word into individual letters}
\usage{
parse_word(word)
}
\arguments{
\item{word}{single character string}
}
\value{
vector of length-1 characters
}
\description{
Split word into individual letters
}
\examples{
parse_word("word")
}
