% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordle-game.R
\name{play_wordle}
\alias{play_wordle}
\title{Play Wordle in the R console
This is a very naive implementation.}
\usage{
play_wordle(
  words = wordle_dict,
  dark_mode = TRUE,
  debug = FALSE,
  target_word = NULL
)
}
\arguments{
\item{words}{character vector of candidate words}

\item{dark_mode}{if using darkmode, set to TRUE}

\item{debug}{logical. default FALSE}

\item{target_word}{specify the target word instead of choosing randomly.
Useful for debugging.}
}
\description{
Play Wordle in the R console
This is a very naive implementation.
}
