% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_keyboard}
\alias{plot_keyboard}
\title{Plot keyboard grid as hint}
\usage{
plot_keyboard(df)
}
\arguments{
\item{df}{Data frame built by `build_keyboard_df`}
}
\value{
ggplot
}
\description{
Plot keyboard grid as hint
}
