% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wordle.R
\name{plot_wordle}
\alias{plot_wordle}
\title{Plotting your guesses and remaining keys}
\usage{
plot_wordle(x, y)
}
\arguments{
\item{x}{a data.frame}

\item{y}{a data.frame}
}
\value{
a \code{ggplot} object
}
\description{
Plotting your guesses and remaining keys
}
\examples{
\dontrun{
plot_wordle(x, y)
}
}
\seealso{
\code{\link{wordle}}
}
