% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordle.R
\name{process_word}
\alias{process_word}
\title{Process filters for a given word and its colors}
\usage{
process_word(words, word, colors)
}
\arguments{
\item{words}{Vector of words to start with}

\item{word}{Current word to process}

\item{colors}{Current colors to process}
}
\value{
Output words vector
}
\description{
Process filters for a given word and its colors
}
