% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_wordle.R
\name{prompt_for_guess}
\alias{prompt_for_guess}
\title{Request word guess input from user}
\usage{
prompt_for_guess(n_letters = 5, init = TRUE)
}
\arguments{
\item{n_letters}{Required number of letters in guess word}

\item{init}{Whether this is the initial prompt of the Wordle game.
Determines which output to display to user when requesting guess.}
}
\value{
input word
}
\description{
Request word guess input from user
}
