% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordle.R
\name{show_words}
\alias{show_words}
\title{Show possible wordle words v.2}
\usage{
show_words(
  word_vec = character(0),
  color_vec = character(0),
  master_wordlist = wordle::words
)
}
\arguments{
\item{word_vec}{Ordered character vector of guessed words}

\item{color_vec}{Ordered character vector of colors of guessed words}

\item{master_wordlist}{Master word list to use, use default unless testing}
}
\value{
A vector of possible words
}
\description{
Show possible wordle words v.2
}
