% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word-scoring.R
\name{sort_by_english_letter_freq}
\alias{sort_by_english_letter_freq}
\title{Sort a list of words with words containing the most common english letters coming first}
\usage{
sort_by_english_letter_freq(words)
}
\arguments{
\item{words}{character vector}
}
\value{
Sorted, named numeric vector mapping words to letter freq scores
}
\description{
Sort a list of words with words containing the most common english letters coming first
}
