% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word-scoring.R
\name{sort_by_letter_matches}
\alias{sort_by_letter_matches}
\title{Calculate mean number of matching letters between each word and every other word}
\usage{
sort_by_letter_matches(words)
}
\arguments{
\item{words}{character vector}
}
\value{
named numeric vector. Names = word, values = score
}
\description{
Calculate mean number of matching letters between each word and every other word
}
