% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word-scoring.R
\name{sort_by_letter_position_correspondence}
\alias{sort_by_letter_position_correspondence}
\title{Calculate mean match for identical letter positions between each word and every other word}
\usage{
sort_by_letter_position_correspondence(words, quiet = FALSE)
}
\arguments{
\item{words}{character vector}

\item{quiet}{suppress output? default: FALSE}
}
\value{
named numeric vector. Names = word, values = score
}
\description{
Calculate mean match for identical letter positions between each word and every other word
}
