% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{target_dictionary}
\alias{target_dictionary}
\title{Wordle dictionary}
\format{
Character vector
}
\source{
\url{https://www.powerlanguage.co.uk/}
}
\usage{
target_dictionary
}
\description{
2,315 5-letter words used as possible target words when playing Wordle
}
\keyword{datasets}
