% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_wordle.R
\name{validate_guess}
\alias{validate_guess}
\title{Validate a Wordle guess}
\usage{
validate_guess(guess, n_letters = 5)
}
\arguments{
\item{guess}{Guessed word}

\item{n_letters}{Required number of letters for the guess}
}
\value{
List of logical values for each word check
}
\description{
Validate a Wordle guess
}
