% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vectorise_word}
\alias{vectorise_word}
\title{Vectorise a word}
\usage{
vectorise_word(word)
}
\arguments{
\item{word}{a word}
}
\value{
A character vector with each letter of the word as an element
}
\description{
Vectorise a word
}
\examples{
vectorise("apple")
}
