% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word-scoring.R
\name{word2pos}
\alias{word2pos}
\title{Convert a word to a integer vector indicating which letters are present}
\usage{
word2pos(word)
}
\arguments{
\item{word}{string}
}
\value{
integer vector representing the alphabet with letters marked with their
        position within the word
}
\description{
Convert a word to a integer vector indicating which letters are present
}
