% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word-scoring.R
\name{word2vec}
\alias{word2vec}
\title{Convert a word to a logical vector indicating which letters are present}
\usage{
word2vec(word)
}
\arguments{
\item{word}{string}
}
\value{
logical vector
}
\description{
Convert a word to a logical vector indicating which letters are present
}
