% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_bank.R
\docType{data}
\name{word_bank}
\alias{word_bank}
\title{Vector of all 5 letter words
Combined two word lists gathered from GitHub user *cfreshman*.}
\format{
An object of class \code{character} of length 12972.
}
\source{
\url{https://gist.github.com/cfreshman}
}
\usage{
word_bank
}
\description{
Vector of all 5 letter words
Combined two word lists gathered from GitHub user *cfreshman*.
}
\keyword{datasets}
