% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordle-package.R
\docType{package}
\name{wordle-package}
\alias{wordle}
\alias{wordle-package}
\title{wordle: Play Wordle (Optionally Automatically)}
\description{
Play Wordle with automation.
}
\author{
\strong{Maintainer}: Bill Denney \email{wdenney@humanpredictions.com} (\href{https://orcid.org/0000-0002-5759-428X}{ORCID})

}
\keyword{internal}
