% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordle.R
\name{Wordle}
\alias{Wordle}
\title{R6 Class for tracking Wordle game state}
\description{
R6 Class for tracking Wordle game state

R6 Class for tracking Wordle game state
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{words}}{current list of candidated words which match all information
so far}

\item{\code{exact}}{string representing which letters are known exactly, with
a period used to indicate that the letter at this position is
unknown}

\item{\code{excluded_letters}}{all letters which are known to not be in the
target word. character vector}

\item{\code{wrong_spot}}{character vector of strings. Each string contains all
letters which are known to be part of the word, but do not exist
at this particular position}

\item{\code{known_count}}{named vector of known counts for letters e.g. if
'e' is known only to appear once, then `known_count = c(e = 1)`}

\item{\code{nchar}}{number of characters in word
Initialize Wordle}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Wordle$new()}}
\item \href{#method-get_suggestions}{\code{Wordle$get_suggestions()}}
\item \href{#method-update}{\code{Wordle$update()}}
\item \href{#method-clone}{\code{Wordle$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wordle$new(nchar, word_file = "/usr/share/dict/words")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nchar}}{number of characters in the word}

\item{\code{word_file}}{source for all words. text file with 1-line-per-word
Get some suggestions on candidate words}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_suggestions"></a>}}
\if{latex}{\out{\hypertarget{method-get_suggestions}{}}}
\subsection{Method \code{get_suggestions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wordle$get_suggestions(n = 20)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of words. default 20
Update the internal list of candidate words}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wordle$update(word, response)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{word}}{the attempted word that was entered into the system}

\item{\code{response}}{a vector of colours representing the response from the
system to this word. Allowed colours: "green", "yellow", "grey"
or "gray"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wordle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
