% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_bank.R
\docType{data}
\name{wordle_allowed}
\alias{wordle_allowed}
\title{Vector of all 5 letter allowed words}
\format{
An object of class \code{character} of length 10657.
}
\source{
\url{https://gist.github.com/cfreshman/cdcdf777450c5b5301e439061d29694c}
}
\usage{
wordle_allowed
}
\description{
All wordle words that are allowed by wordle.
}
\keyword{datasets}
