% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_bank.R
\docType{data}
\name{wordle_answers}
\alias{wordle_answers}
\title{Vector of all 5 letter answer words}
\format{
An object of class \code{character} of length 2315.
}
\source{
\url{https://gist.github.com/cfreshman/a03ef2cba789d8cf00c08f767e0fad7b}
}
\usage{
wordle_answers
}
\description{
All wordle words that could be the solution.
}
\keyword{datasets}
