% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplay.R
\name{wordle_autoplay}
\alias{wordle_autoplay}
\title{Automatically play a Wordle game}
\usage{
wordle_autoplay(
  game_state,
  guess_fun,
  initial_guess,
  verbose_guess = FALSE,
  verbose_state = FALSE
)
}
\arguments{
\item{game_state}{The initial game state}

\item{guess_fun}{A guessing function (preferably cached), see
\code{?wordle_choose_guess} or \code{?wordle_choose_guess_cache}}

\item{initial_guess}{The first word to guess}

\item{verbose_guess, verbose_state}{Be verbose for guessing or state
generation?}
}
\value{
The final game state
}
\description{
Automatically play a Wordle game
}
