% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessing.R
\name{wordle_choose_guess}
\alias{wordle_choose_guess}
\title{Choose a guess based on a game_state and a metric function}
\usage{
wordle_choose_guess(game_state, metric_fun, verbose = TRUE)
}
\arguments{
\item{game_state}{The output of \code{new_wordle_game()} or \code{wordle_game_state()}}

\item{metric_fun}{The metric function to use for guessing; the guess with the
lowest metric value will be used, and being in the possible words will be
used if more than one has an equal metric value.}

\item{verbose}{Show more information}
}
\value{
The guess
}
\description{
Choose a guess based on a game_state and a metric function
}
\seealso{
Other Wordle guessing: 
\code{\link{wordle_choose_guess_cache}()}
}
\concept{Wordle guessing}
