% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessing.R
\name{wordle_choose_guess_cache}
\alias{wordle_choose_guess_cache}
\title{Generate a guessing cache function}
\usage{
wordle_choose_guess_cache(metric_fun)
}
\arguments{
\item{metric_fun}{The function to use for calculation of the metric (see
\code{?wordle_metric})}
}
\value{
A function with a cache
}
\description{
Generate a guessing cache function
}
\seealso{
Other Wordle guessing: 
\code{\link{wordle_choose_guess}()}
}
\concept{Wordle guessing}
