% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wordle_colors}
\alias{wordle_colors}
\title{"Colors" for Wordle guesses}
\format{
A data.frame with rownames for the possible words (from
\code{wordlists$possible}), the first column named "X" with the words
repeated, and columns for each of the possible guesses. The values after
the first column are integers indicating the colors of the word given the
guess.  The colors are a 1 to 5-digit integer where 2 indicates that the
letter matches in the current position, 1 indicates that the letter matches
in a different position, and 0 indicates that the letter does not match.
}
\usage{
wordle_colors
}
\description{
"Colors" for Wordle guesses
}
\keyword{datasets}
