% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/words.R
\docType{data}
\name{wordle_dict}
\alias{wordle_dict}
\title{An 'official' wordlist scraped from the Wordle website}
\format{
An object of class \code{character} of length 12972.
}
\usage{
wordle_dict
}
\description{
An 'official' wordlist scraped from the Wordle website
}
\keyword{datasets}
