% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_state.R
\name{wordle_game_state}
\alias{wordle_game_state}
\title{Update the game state with a new guess}
\usage{
wordle_game_state(game_state, guess, verbose = TRUE)
}
\arguments{
\item{game_state}{The game state either as returned by
\code{new_wordle_game()} or this function.}

\item{guess}{The new word to guess}

\item{verbose}{Show progress information}
}
\value{
An updated \code{game_state}
}
\description{
Update the game state with a new guess
}
