% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_setup.R
\name{wordle_guess_colors}
\alias{wordle_guess_colors}
\title{Generate the "colors" for a word guess relative to the possible words}
\usage{
wordle_guess_colors(current_guess, wordle_has_letter_df)
}
\arguments{
\item{guess}{The word that was guessed}

\item{possible}{The vector of possible words}
}
\value{
An integer vector where 0 indicates no in the word, 1 indicates
elsewhere in the word, and 2 indicates at the current location in the word.
}
\description{
Generate the "colors" for a word guess relative to the possible words
}
