% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_setup.R
\name{wordle_guess_colors_df}
\alias{wordle_guess_colors_df}
\title{Generate a data.frame of "colors" where the first column are the possible
words to guess and the remaining columns are the colors generated by the
guessed word.}
\usage{
wordle_guess_colors_df(possible, all_words)
}
\arguments{
\item{possible}{The vector of possible words}
}
\description{
Generate a data.frame of "colors" where the first column are the possible
words to guess and the remaining columns are the colors generated by the
guessed word.
}
