% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{wordle_metric}
\alias{wordle_metric}
\alias{wordle_metric_smallest_big_group}
\alias{wordle_metric_most_to_median_group}
\alias{wordle_metric_smallest_median_group}
\title{Calculation metrics for Wordle}
\usage{
wordle_metric_smallest_big_group(x)

wordle_metric_most_to_median_group(x)

wordle_metric_smallest_median_group(x)
}
\arguments{
\item{x}{A vector of possible colors (see \code{wordle_colors})}
}
\value{
The metric calculated by the guessing method (lower values are
better) as a named numeric vector where the names are the guess and the
value is the metric value.
}
\description{
Calculation metrics for Wordle
}
\section{Functions}{
\itemize{
\item \code{wordle_metric_smallest_big_group}: Choose the largest group size

\item \code{wordle_metric_most_to_median_group}: Choose the largest number to the median word

\item \code{wordle_metric_smallest_median_group}: Choose the smallest sized median group
}}

