% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/words-solutions.R
\docType{data}
\name{wordle_solns}
\alias{wordle_solns}
\title{Solutions to the wordle puzzles embedded in the JS}
\format{
An object of class \code{character} of length 2315.
}
\usage{
wordle_solns
}
\description{
Solutions to the wordle puzzles embedded in the JS
}
\keyword{datasets}
