% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordle-game.R
\name{WordleGame}
\alias{WordleGame}
\title{Wordle Game Engine}
\description{
Wordle Game Engine

Wordle Game Engine
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{words}}{character vector of words}

\item{\code{target_word}}{target word chosen random during initialisation}

\item{\code{target}}{character vector of letters in target word}

\item{\code{attempts}}{vector of attempted words}

\item{\code{responses}}{list of responses for each attempte}

\item{\code{nchar}}{number of characters in all words}

\item{\code{dark_mode}}{logical.
Initialise a WordleGame object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{WordleGame$new()}}
\item \href{#method-try}{\code{WordleGame$try()}}
\item \href{#method-is_solved}{\code{WordleGame$is_solved()}}
\item \href{#method-share}{\code{WordleGame$share()}}
\item \href{#method-clone}{\code{WordleGame$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WordleGame$new(words, dark_mode = TRUE, debug = FALSE, target_word = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{words}}{character vector of candidate words}

\item{\code{dark_mode}}{if using darkmode, set to TRUE}

\item{\code{debug}}{logical. default FALSE}

\item{\code{target_word}}{specify the target word instead of choosing randomly.
       Useful for debugging.
Play a word and see what the response is}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-try"></a>}}
\if{latex}{\out{\hypertarget{method-try}{}}}
\subsection{Method \code{try()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WordleGame$try(word)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{word}}{string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_solved"></a>}}
\if{latex}{\out{\hypertarget{method-is_solved}{}}}
\subsection{Method \code{is_solved()}}{
is the game solved?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WordleGame$is_solved()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-share"></a>}}
\if{latex}{\out{\hypertarget{method-share}{}}}
\subsection{Method \code{share()}}{
Print a shareable text block showing the attempts
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WordleGame$share()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WordleGame$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
