% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordle-helper.R
\name{WordleHelper}
\alias{WordleHelper}
\title{R6 Class for tracking Wordle game state}
\description{
R6 Class for tracking Wordle game state

R6 Class for tracking Wordle game state
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{words}}{current list of candidated words which match all information
so far}

\item{\code{exact}}{string representing which letters are known exactly, with
a period used to indicate that the letter at this position is
unknown}

\item{\code{wrong_spot}}{character vector of strings. Each string contains all
letters which are known to be part of the word, but do not exist
at this particular position}

\item{\code{min_count}}{named vector of known minimum counts}

\item{\code{known_count}}{named vector of known counts for letters e.g. if
'e' is known only to appear once, then `known_count = c(e = 1)`}

\item{\code{nchar}}{number of characters in word
Initialize Wordle}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{WordleHelper$new()}}
\item \href{#method-update}{\code{WordleHelper$update()}}
\item \href{#method-clone}{\code{WordleHelper$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WordleHelper$new(nchar, words = wordle_dict)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nchar}}{number of characters in the word}

\item{\code{words}}{character vector of candidated words
Update the internal list of candidate words}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WordleHelper$update(word, response)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{word}}{the attempted word that was entered into the system}

\item{\code{response}}{a vector of colours representing the response from the
system to this word. Allowed colours: "green", "yellow", "grey"
or "gray"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WordleHelper$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
