% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wordlists}
\alias{wordlists}
\title{Words for guessing and that may be correct in Wordle}
\format{
A list with two named elements:
\describe{
\item{possible}{Words that are possibly correct}
\item{all}{Words that will not be correct, but can be guessed}
}
}
\usage{
wordlists
}
\description{
Words for guessing and that may be correct in Wordle
}
\keyword{datasets}
