% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{words}
\alias{words}
\title{5-letter English words in UPPERCASE}
\format{
\subsection{\code{words}}{

A character vector of length 15918:
}
}
\source{
\url{https://github.com/dwyl/english-words/blob/master/words_alpha.txt}
}
\usage{
words
}
\description{
5-letter English words in UPPERCASE
}
\keyword{datasets}
