% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordle.R
\name{filter_grays}
\alias{filter_grays}
\title{Filter word list to exclude those with gray letters}
\usage{
filter_grays(words, grays = "")
}
\arguments{
\item{words}{A character vector of words to filter.}

\item{grays}{A string containing gray wordle letters, e.g. \code{"ABOPD"}}
}
\value{
A character vector of filtered words
}
\description{
Filter word list to exclude those with gray letters
}
