% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordle.R
\name{filter_greens}
\alias{filter_greens}
\title{Filter word list to only include those that match pattern in greens}
\usage{
filter_greens(words, greens)
}
\arguments{
\item{words}{A character vector of words to filter.}

\item{greens}{A single string, e.g. \code{"A--I-"}. Greens are letters, and hyphens are placeholders for non-greens.}
}
\value{
A character vector of filtered words
}
\description{
Filter word list to only include those that match pattern in greens
}
