% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordle.R
\name{filter_yellows}
\alias{filter_yellows}
\title{Filter word list to those consistent with yellows}
\usage{
filter_yellows(words, yellows = rep("", 5))
}
\arguments{
\item{words}{A character vector of words to filter.}

\item{yellows}{A character vector of length 5. Each element is a string of yellows at that position. If no yellows at a position, string should be "".}
}
\value{
A character vector of filtered words
}
\description{
Filter word list to those consistent with yellows
}
