% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_words.R
\name{find_words}
\alias{find_words}
\title{Filter dictionary using regular expression}
\usage{
find_words(dict, pattern, negate = FALSE)
}
\arguments{
\item{dict}{Character vector through which to search for pattern}

\item{pattern}{Regular expression to match with entries in \code{dict}}

\item{negate}{If \code{TRUE}, include strings in \code{dict} that match with \code{pattern}. Otherwise, exclude matches.}
}
\value{
Character vector
}
\description{
Filter dictionary using regular expression
}
\examples{
dictionary <- c("this", "is", "a", "dictionary")
find_words(dictionary, pattern = "^d")
}
