% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_letters.R
\name{make_guess}
\alias{make_guess}
\title{Make a single Wordle guess}
\usage{
make_guess(target, guess)
}
\arguments{
\item{target}{target of the match}

\item{guess}{word to be compared to the target word}
}
\value{
List of length 3.
The first element is a list of lists, each of which contains a letter in the
guess word that was present in the target word and its associated position in the word.
The second element is a character vector of letters in the guess not present in the target word.
The third element is a list of lists, each of which contains a misplaced letter and its associated
position in the guess
}
\description{
Compare a guess word to a target word and return correct,
incorrect, and misplaced letters with associated index positions.
}
\examples{
guess <- make_guess("torch", "trick")
str(guess)
}
