% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_wordle.R
\name{make_guess}
\alias{make_guess}
\title{Make a single Wordle guess}
\usage{
make_guess(target, guess, quiet = FALSE)
}
\arguments{
\item{target}{Target of the match}

\item{guess}{Word to be compared to the target word}

\item{quiet}{Should guess results be printed to console in the form of color-
coded guess letters?}
}
\value{
List of length 3. Each list element is a vector of integers indicating
index positions for the letters in the guess word in the associated category.
}
\description{
Compare a guess word to a target word and return index positions
of correct, incorrect, and misplaced letters. Optionally print guessed
letters to console with colors indicating whether letters were correct,
incorrect, or misplaced.
}
