% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_wordle.R
\name{play_wordle}
\alias{play_wordle}
\title{Play a game of Wordle}
\usage{
play_wordle(n_letters = 5)
}
\arguments{
\item{n_letters}{Length of target word}
}
\description{
Selects a random target word of specified length and prompts user
for input words. Input words are compared to target word and feedback
is given to user indicating which words are correct, incorrect, or
misplaced.
}
