% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordle.R
\name{show_words}
\alias{show_words}
\title{Show possible wordle words}
\usage{
show_words(grays = "", yellows = rep("", 5), greens = "-----")
}
\arguments{
\item{grays}{grays, see \code{\link[=filter_grays]{filter_grays()}}}

\item{yellows}{yellows, see \code{\link[=filter_yellows]{filter_yellows()}}}

\item{greens}{greens, see \code{\link[=filter_greens]{filter_greens()}}}
}
\value{
A vector of possible words
}
\description{
Show possible wordle words
}
