% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_bank.R
\docType{data}
\name{word_bank}
\alias{word_bank}
\title{Vector of 5 letter words}
\format{
An object of class \code{character} of length 9330.
}
\source{
\url{https://github.com/cran/words}
}
\usage{
word_bank
}
\description{
Vector that was obtained and filtered from the {words} package.
}
\keyword{datasets}
