% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordle.R
\name{wordle}
\alias{wordle}
\title{Play wordle with your R package names.}
\usage{
wordle(answer = "installed", strict = TRUE)
}
\arguments{
\item{answer}{\code{"installed"}, \code{"cran"}, \code{"ranking"}, or an atomic vector of any five character.}

\item{strict}{a logical value (default is \code{FALSE})}
}
\description{
Play wordle with your R package names.
}
\details{
\describe{
\item{\code{answer}}{If \code{"installed"}, a package list is obtained from your PC. If \code{"cran"}, a package list is obtained from CRAN (extremely hard mode!). If \code{"ranking"}, a top downloaded package list is obtained from CRAN mirror using \{cranlogs\} package. Also, you can set any answer with five characters.}
\item{\code{strict}}{If \code{TRUE}, your guess not in package list is not applicable. If \code{FALSE} (default), your guess can have any five characters.}
}
}
\examples{
\dontrun{
wordle()
}
}
\seealso{
\url{https://www.nytimes.com/games/wordle/index.html}
}
