% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wordle.R
\name{wordle}
\alias{wordle}
\title{wordle}
\usage{
wordle(answer = "installed", strict = FALSE)
}
\arguments{
\item{answer}{\code{"installed"}, \code{"cran"}, or any five character.}

\item{strict}{a logical value (default is \code{FALSE})}
}
\description{
wordle
}
\details{
\describe{
\item{\code{answer}}{If \code{"installed"}, a pacakge list is obtained from your PC. If \code{"cran"}, a package list is obtained from CRAN. Also, you can set any answer with five characters.}
\item{\code{strict}}{If \code{TRUE}, your guess not in package list is not applicable (default is \code{FALSE}).}
}
}
